Upload of initialization code
These lines of code change the startup mode of the VL6180X micro-LIDAR sensor to threshold interrupt mode where the sensor toggles the voltage on it's INT pin when something gets too close to the sensor relative to specified distance. This pin can then be tied to an interrupt on the Arduino to cause the Arduino to react.